+++ R/gtrends.R
@@ -7,7 +7,7 @@
 #' If the environment variables \code{GOOGLE_USER} and
 #' \code{GOOGLE_PASSWORD} are set, they will be retrieved in case no
 #' argument has been supplied.  Similarly, the environment variable
-#' \code{GOOGLE_PASSWORD} or \code{options("google.password")} can be
+#' \code{options("google.user")} or \code{options("google.password")} can be
 #' used. Lastly, if the environment variable \code{GOOGLE_AUTOCONNECT}
 #' is set to (the text string) \sQuote{TRUE}, or the the R option
 #' \code{options("google.autoconnect")} is set to \sQuote{TRUE} then
@@ -148,16 +148,16 @@ gconnect <- function(usr = NULL, psw = NULL, verbose = FALSE) {
 #'   Multiple keywords are possible using \code{gtrends(c("NHL", "NBA", "MLB", 
 #'   "MLS"))}.
 #'   
-#' @param geo A character vector denoting geographic regions for the query,
-#'   default to \dQuote{all} for global queries. Multiple regions are possible
+#' @param geo A character vector denoting geographic regions for the query, 
+#'   default to \dQuote{all} for global queries. Multiple regions are possible 
 #'   using \code{gtrends("NHL", c("CA", "US"))}.
 #'   
 #' @param cat A character denoting the category, defaults to \dQuote{0}.
-#'   
-#' @param ... Additional parameters passed on in method dispatch.
-#'   
-#' @param res Resolution of the trending data to be returned. Either \code{week}
-#'   for weekly data or \code{day} for daily data.
+#'      
+#' @param res Resolution of the trending data to be returned. One of 
+#'   \code{c("1h", "4h", "1d", "7d")}. If \code{res} is provided, then 
+#'   \code{start_date} and \code{end_date} parameters are ignored. See 
+#'   \emph{Query resolution} for more information.
 #'   
 #' @param start_date Starting date using yyyy-mm-dd format. Must be greater than
 #'   2004-01-01.
@@ -169,14 +169,47 @@ gconnect <- function(usr = NULL, psw = NULL, verbose = FALSE) {
 #'   Users can either supply an explicit handle, or rely on the helper function 
 #'   \code{.getDefaultConnection()} to retrieve the current connection handle.
 #'   
+#' @param ... Additional parameters passed on in method dispatch.
+#'
+#' @section Query resolution: By default, Google returns weekly information when
+#'   the requested data spans a period greater than three months. It is also 
+#'   possible to obtain \emph{daily} and \emph{hourly} information. However, 
+#'   these are only available for a certain period prior to the \emph{current} 
+#'   date.
+#'   
+#'   For instance, \code{1h}, \code{7h}, \code{1d} and \code{7d} denote 
+#'   trends data for the last 1 hour, last four hours, last day and last seven 
+#'   days respectively. Using one of the above \code{res} will return the 
+#'   corresponding hourly data.
+#'   
+#'   Note that data requested for a beriod between one and three months will be 
+#'   returned daily. For a  period greater than three months, data will be 
+#'   always returned weekly.
+#'   
 #' @return An object of class \sQuote{gtrends} which is list with six elements 
 #'   containing the results.
 #' @examples 
 #' \dontrun{
 #' ch <- gconnect("usr@gmail.com", "psw")
-#' sport_trend <- gtrends(c("NHL", "NBA", "MLB", "MLS"))
 #' 
-#' sport_trend <- gtrends("NHL", geo = c("CA", "US"))
+#' gtrends(c("NHL", "NBA", "MLB", "MLS"))
+#' 
+#' gtrends("NHL", geo = c("CA", "US"))
+#' 
+#' # Search only for the sport category.
+#' gtrends("NHL", geo = c("CA", "US"), cat = "0-20")
+#' 
+#' # Trends between 2015-01-01 and 2015-03-01 in Sweeden. Will be daily data.
+#' gtrends("NHL", geo = c("SE"), start_date = "2015-01-01", end_date = "2015-03-01")
+#' 
+#' # Trends between 2015-01-01 and 2015-04-01 in Sweeden. Will be weekly data.
+#' gtrends("NHL", geo = c("SE"), start_date = "2015-01-01", end_date = "2015-04-01")
+#' 
+#' # Last 4 hours trends
+#' gtrends("NHL", geo = c("CA"), res = "4h")
+#' 
+#' # Last 7 days trends
+#' gtrends("NHL", geo = c("CA"), res = "7d")
 #' }
 #' @export
 gtrends <- function(query, geo, cat, ch, ...) {
@@ -192,7 +225,7 @@ gtrends.default <- function(query,
                             geo, 
                             cat, 
                             ch, 
-                            res = "week",
+                            res = c(NA, "1h", "4h", "1d", "7d"),
                             start_date = as.Date("2004-01-01"),
                             end_date = as.Date(Sys.time()),
                             ...){
@@ -203,10 +236,10 @@ gtrends.default <- function(query,
 
   stopifnot(is.character(query),
             is.vector(query),
-            all(res %in% c("week", "day")),
-            length(res) == 1,
             length(query) <= 5, 
             length(geo) <= 5)
+  
+  res <- match.arg(res, several.ok = FALSE)
   
   if(length(query) > 1 & length(geo) > 1){
     stop("Can not specify multiple keywords and geo at the same time.", 
@@ -223,40 +256,54 @@ gtrends.default <- function(query,
          call. = FALSE)
   }
   
-  ## Verify the dates
+  #---------------------------------------------------------------------
+  # Date verification.
+  #---------------------------------------------------------------------
+  
   start_date <- as.Date(start_date, "%Y-%m-%d")  
   end_date <- as.Date(end_date, "%Y-%m-%d")  
   
   if (is.na(start_date)) {
-    stop("start_date is not a valide date. Please use yyyy-mm-dd format.",
+    stop("start_date is not a valid date. Please use yyyy-mm-dd format.",
          call. = FALSE)
   } 
   
   if (is.na(end_date)) {
-    stop("end_date is not a valide date. Please use yyyy-mm-dd format.",
+    stop("end_date is not a valid date. Please use yyyy-mm-dd format.",
          call. = FALSE)
   } 
   
-  # date verification
   stopifnot(start_date < end_date, 
             start_date >= as.Date("2004-01-01"), # cant be earlier than 2004
             end_date <= as.Date(Sys.time())) # cant be more than current date
   
-  # if resolution is day then maximum date difference must be less than 3 months
   nmonth <- length(seq(from = start_date, to = end_date, by = "month"))
   
-  if(res == "day" & nmonth > 3){
-    stop("Maximum of 3 months allowed with daily resolution.", call. = FALSE)
+  if(nmonth >= 1){
+    date <- paste(format(start_date, "%m/%Y"), paste(nmonth, "m", sep = ""))
   }
   
+  if(!is.na(res)){
+    
+    # Match Google code (ex. 1-H) to a more user friendly value (1h)
+    resolution_code <- data.frame(code = c("1-H", "4-H", "1-d", "7-d"),
+                                  res = c("1h", "4h", "1d", "7d"), 
+                                  stringsAsFactors = FALSE)
+    
+    res <- resolution_code$code[resolution_code$res == res]
+    
+    date <- paste("now" , res)
+  }
+
+  #---------------------------------------------------------------------
+  # Build the query.
+  #---------------------------------------------------------------------
   query <- paste(query, collapse = ",")
   
   ## Change encoding to utf-8
   if (!(Encoding(query) == "UTF-8")) {
     query <- iconv(query, "latin1", "utf-8", sub = "byte")
   }
-  
-  data(countries, envir = environment())
   
   countries[, 1] <- as.character(countries[, 1])
   countries[, 2] <- as.character(countries[, 2])
@@ -267,8 +314,6 @@ gtrends.default <- function(query,
          call. = FALSE)
   }
   
-  # https://www.google.com/trends/trendsReport?&q=nhl&geo=US%2C%20BR&cmpt=geo&content=1&export=1
-  
   geo <- paste(geo, sep = "", collapse = ", ")
   #geo <- URLencode(geo, reserved = TRUE)
   
@@ -276,26 +321,20 @@ gtrends.default <- function(query,
   
   trendsURL <- "http://www.google.com/trends/trendsReport?"
 
-  
-  res <- paste(nmonth, "m", sep = "")
-  
   pp <- list(q = query, 
+             cat = cat,
              cmpt = cmpt, 
              content = 1, 
              export = 1,
-             date = paste(format(start_date, "%m/%Y"), res),
+             date = date,
              geo = geo)
-  
-  
-  #http://www.google.com/trends/trendsReport?&q=%2Cfoo%2Cbar%2Cbaz%2Cfoo&cmpt=q&content=1&export=1&date=1%2F2015%202m
   
   resultsText <- getForm(trendsURL, .params = pp, curl = ch)
   
   if (any(grep("QUOTA", resultsText))) {
-  
     stop("Reached Google Trends quota limit! Please try again later.")
-  
   }
+  
   queryparams <- c(query = query, 
                    cat = cat, 
                    geo = geo, 
@@ -338,6 +377,8 @@ summary.gtrends <- function(object, ...) {
 #' @param which Block number containing the geographical data to plot.
 #' @param ind A integer selecting the result set in case of multiple
 #' search terms.
+#' @return When \code{type} is equal to \sQuote{trends}, the resulting 
+#' ggplot2 object is returned silently.
 #' @import googleVis
 #' @import ggplot2
 #' @importFrom graphics plot
@@ -350,6 +391,7 @@ summary.gtrends <- function(object, ...) {
 plot.gtrends <- function(x, type = c("trend", "geo"), which = 5, ind = 1L, ...){
   
   type <- match.arg(type)
+  ret <- NULL # by default we return nothing
   
   if (type == "trend") {
     
@@ -373,6 +415,7 @@ plot.gtrends <- function(x, type = c("trend", "geo"), which = 5, ind = 1L, ...){
       theme_bw()
     
     print(p)
+    ret <- p
     
   } else if (type == "geo") {
     
@@ -383,8 +426,6 @@ plot.gtrends <- function(x, type = c("trend", "geo"), which = 5, ind = 1L, ...){
     block <- x[which][[ind]]
     
     # Try to find if the requested block contains geographic information.
-    data(locations, envir = environment())
-    
     if(!any(tolower(block[1, ]) %in% tolower(locations$Name))){
       
       message("The requested block does not seems to contain geographical information. Please choose another block.")
@@ -408,8 +449,7 @@ plot.gtrends <- function(x, type = c("trend", "geo"), which = 5, ind = 1L, ...){
                                      resolution = "countries")))
   } 
   
-  
-  invisible(NULL)
+  invisible(ret)
 }
 
 #' @rdname gtrends
@@ -436,51 +476,61 @@ as.zoo.gtrends <- function(x, ...) {
          call. = FALSE)
   }
   
-  ## block 1: meta data
-  meta  <- strsplit(vec[1], "\\\r\\\n")[[1]]
-
   #---------------------------------------------------------------------
   # Section to deal with trend data.
   #---------------------------------------------------------------------
-    
-  ## block 2: trend
+  
+  # meta data
+  meta  <- strsplit(vec[1], "\\\r\\\n")[[1]]
+
+  # trend
   trend <- read.csv(textConnection(strsplit(vec[2], "\\\n")[[1]]),
                     skip = 1,
                     stringsAsFactors = FALSE)
   
-  weeks <- data.frame(do.call(rbind, strsplit(trend[, 1], " - ")))
+  # block date
+  
+  weeks <- data.frame(date = do.call(rbind, strsplit(trend[, 1], " - ")),
+                      stringsAsFactors = FALSE)
   
   trend <- trend[, mapply(is.numeric, trend), drop = FALSE]
   
-  names(trend) <- unlist(strsplit(queryparams[1], ","), use.names = FALSE)
+  #names(trend) <- unlist(strsplit(queryparams[1], ","), use.names = FALSE)
   
-  is_weekly <- all(do.call(c, 
-                           lapply(weeks, 
-                                  grepl, 
-                                  pattern = "\\d{4}-\\d{2}-\\d{2}")))
+  kw <- trimws(unlist(strsplit(queryparams[1], ","), use.names = FALSE))
+  geo <- trimws(unlist(strsplit(queryparams[3], ","), use.names = FALSE))
+  names(trend) <- make.names(paste(kw, geo))
   
-  is_daily <- all(do.call(c, 
-                           lapply(weeks, 
-                                  grepl, 
-                                  pattern = "\\d{4}-\\d{2}")))
   
-  # data = gtrends(c("nhl", "asasnfassaasl"), "IS") # Bug here with IS
-  if(!is_weekly & !is_daily){
-    stop("Not enough search volume to show results.", call. = FALSE)
-  }
-  
-  if(is_weekly){
+  if(ncol(weeks) == 2){
     
-    weeks <- lapply(weeks, as.Date, SIMPLIFY = FALSE)
+    weeks <- lapply(weeks, as.POSIXct, SIMPLIFY = FALSE)
     weeks <- do.call(cbind.data.frame, weeks)
     names(weeks) <- c("start", "end")[1:ncol(weeks)]
   
-  }else{
+  }
+  
+  # Either daily or hourly data
+  if(ncol(weeks) == 1){
     
-    weeks <- paste(weeks[, 1], "-01", sep = "")
-    weeks <- data.frame(start = weeks)
+    if(nchar(weeks$date[1]) == 7){
+      
+      # Sometimes data are returned without a day. Asusme it is first day of month.
+      weeks <- as.POSIXct(paste(weeks[, 1], "-01", sep = ""))
+      weeks <- data.frame(start = weeks)
+      
+    }else if(nchar(weeks$date[1]) == 10){
+      
+      weeks <- as.POSIXct(weeks$date)
+      weeks <- data.frame(start = weeks)
+      
+    }else{
+      
+      weeks <- as.POSIXct(weeks[, 1], format = "%Y-%m-%d-%H:%M", tz = "UTC")
+      weeks <- data.frame(start = weeks)
+      
+    }
     
-    message("The number of hits was too low for daily (weekly) resolution. Results were returned using weekly (monthly) resolution instead.")
   }
   
   trend <- cbind(weeks, trend)